+++ backend/src/main/java/br/com/brunodezorzi/API/noticiario/repository/AutorRepository.java
@@ -3,4 +3,4 @@ package br.com.brunodezorzi.api.noticiario.repository;
 import br.com.brunodezorzi.api.noticiario.model.Autor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AutorRepository extends JpaRepository<Autor, Long> {}
+public interface AutorRepository extends JpaRepository<Autor, Integer> {}
