+++ backend/src/main/java/br/com/brunodezorzi/API/noticiario/repository/CategoriaRepository.java
@@ -3,4 +3,5 @@ package br.com.brunodezorzi.api.noticiario.repository;
 import br.com.brunodezorzi.api.noticiario.model.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {}
+public interface CategoriaRepository
+  extends JpaRepository<Categoria, Integer> {}
