+++ backend/src/main/java/br/com/brunodezorzi/API/noticiario/repository/NoticiaRepository.java
@@ -3,4 +3,4 @@ package br.com.brunodezorzi.api.noticiario.repository;
 import br.com.brunodezorzi.api.noticiario.model.Noticia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticiaRepository extends JpaRepository<Noticia, Long> {}
+public interface NoticiaRepository extends JpaRepository<Noticia, Integer> {}
