+++ backend/src/main/java/br/com/brunodezorzi/API/noticiario/repository/PessoaRepository.java
@@ -3,4 +3,4 @@ package br.com.brunodezorzi.api.noticiario.repository;
 import br.com.brunodezorzi.api.noticiario.model.Pessoa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PessoaRepository extends JpaRepository<Pessoa, Long> {}
+public interface PessoaRepository extends JpaRepository<Pessoa, Integer> {}
